Support study.image_received events
Listen for study.image_received events and store the data.
