+++ poc_seven/gateway/Dockerfile.pacs
@@ -20,7 +20,8 @@ RUN mkdir -p /scripts /var/lib/pacs/storage /var/lib/pacs/thumbnails /var/log/pa
 RUN UV_BREAK_SYSTEM_PACKAGES=1 uv pip install --system --no-cache \
     pydicom==3.0.1 \
     pynetdicom==3.0.4 \
-    websockets==14.1
+    websockets==14.1 \
+    Pillow>=12.0
 
 # Copy PACS server scripts
 COPY scripts/pacs_server.py /scripts/
